+++ src/pages/Chat.tsx
@@ -1458,7 +1458,7 @@ export default function Chat({ chatId, isGroup }: { chatId: string; isGroup: boo
             >
               {!isMe && isGroup && (
                 <div className="avatar avatar-sm">
-                  {msg.from_avatar ? <img src={msg.from_avatar} alt="" /> : (msg.from_nickname?.[0] || '?')}
+                  {msg.from_avatar ? <img src={normalizeFileUrl(msg.from_avatar)} alt="" /> : (msg.from_nickname?.[0] || '?')}
                 </div>
               )}
               <div>